pyproject.toml
[tool.poetry]
name = "discordmoderationbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pybooru = "^4.2.2"
"discord.py" = "^1.5.1"
pysaucenao = "^1.5.0"
pygelbooru = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
